backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==6.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
idna==3.11
kombu==5.6.1
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
